Spinbox size is fixed while editing the value with variable precision
Fix the maximum size of the spinbox to the current size when changing the
decimals to maximum possible. This avoids unusual size changes in widgets where
the spinbox is used.
Added new method to set the slider value when the spinbox value is changed
making sure there are no recursive signals fired between the slider and the
spinbox in the ctkSliderWidget.